+++ src/main/java/com/example/minimart/order/repository/OrderItemJpaRepository.java
@@ -1,6 +1,6 @@
-package com.example.minimart.order.infra;
+package com.example.minimart.order.repository;
 
-import com.example.minimart.order.infra.entity.OrderItemEntity;
+import com.example.minimart.order.repository.entity.OrderItemEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
