+++ src/main/java/com/example/minimart/order/repository/OrderJpaRepository.java
@@ -1,6 +1,6 @@
-package com.example.minimart.order.infra;
+package com.example.minimart.order.repository;
 
-import com.example.minimart.order.infra.entity.OrderEntity;
+import com.example.minimart.order.repository.entity.OrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderJpaRepository extends JpaRepository<OrderEntity, Long> {
