+++ src/main/java/com/example/minimart/order/repository/entity/OrderItemEntity.java
@@ -1,45 +1,86 @@
-package com.example.minimart.order.domain;
+package com.example.minimart.order.repository.entity;
+
+import jakarta.persistence.*;
 
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import java.util.Objects;
 
-public class OrderItem {
+@Entity
+@Table(name = "order_items")
+public class OrderItemEntity {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(nullable = false)
+    private Long orderId;
+
+    @Column(nullable = false)
     private Long productId;
+
+    @Column(nullable = false, length = 255)
     private String productName;
+
+    @Column(length = 255)
     private String productOption;
+
+    @Column(nullable = false, precision = 19, scale = 2)
     private BigDecimal unitPrice;
+
+    @Column(nullable = false)
     private int quantity;
+
+    @Column(nullable = false, precision = 19, scale = 2)
     private BigDecimal totalPrice;
 
-    public OrderItem(
-        Long id,
+    @Column(nullable = false, updatable = false)
+    private LocalDateTime createdAt;
+
+    @Column(nullable = false)
+    private LocalDateTime updatedAt;
+
+    @Column(nullable = false)
+    private boolean deleted = false;
+
+    protected OrderItemEntity() {}
+
+    public OrderItemEntity(
+        Long orderId,
         Long productId,
         String productName,
         String productOption,
         BigDecimal unitPrice,
-        int quantity
+        int quantity,
+        BigDecimal totalPrice
     ) {
-        this.id = id;
+        this.orderId = orderId;
         this.productId = productId;
         this.productName = productName;
         this.productOption = productOption;
         this.unitPrice = unitPrice;
         this.quantity = quantity;
-        this.totalPrice = calculateTotalPrice();
+        this.totalPrice = totalPrice;
     }
 
-    public static OrderItem create(Long productId, String productName, String productOption, BigDecimal unitPrice, int quantity) {
-        return new OrderItem(null, productId, productName, productOption, unitPrice, quantity);
+    @PrePersist
+    protected void onCreate() {
+        this.createdAt = LocalDateTime.now();
+        this.updatedAt = LocalDateTime.now();
     }
 
-    private BigDecimal calculateTotalPrice() {
-        return this.unitPrice.multiply(BigDecimal.valueOf(this.quantity));
+    @PreUpdate
+    protected void onUpdate() {
+        this.updatedAt = LocalDateTime.now();
     }
 
     public Long getId() {
         return id;
+    }
+
+    public Long getOrderId() {
+        return orderId;
     }
 
     public Long getProductId() {
@@ -70,7 +111,7 @@ public class OrderItem {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        OrderItem orderItem = (OrderItem) o;
+        OrderItemEntity orderItem = (OrderItemEntity) o;
 
         return Objects.equals(id, orderItem.id);
     }
@@ -84,12 +125,14 @@ public class OrderItem {
     public String toString() {
         return "OrderItem{" +
             "id=" + id +
+            ", orderId=" + orderId +
             ", productId=" + productId +
             ", productName='" + productName + '\'' +
             ", productOption='" + productOption + '\'' +
             ", unitPrice=" + unitPrice +
             ", quantity=" + quantity +
             ", totalPrice=" + totalPrice +
+            ", deleted=" + deleted +
             '}';
     }
