+++ src/main/java/com/example/minimart/order/repository/entity/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.example.minimart.order.infra.entity;
+package com.example.minimart.order.repository.entity;
 
 public enum OrderStatus {
