+++ src/main/java/com/example/minimart/order/service/domain/Order.java
@@ -1,6 +1,6 @@
-package com.example.minimart.order.domain;
+package com.example.minimart.order.service.domain;
 
-import com.example.minimart.order.infra.entity.OrderStatus;
+import com.example.minimart.order.repository.entity.OrderStatus;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
